chore: rename to achieve camelcasing of vMix filenames
