+++ packages/timeline-state-resolver/src/integrations/vmix/__tests__/vMixPollingTimer.spec.ts
@@ -1,4 +1,4 @@
-import { VMixPollingTimer } from '../WMixPollingTimer'
+import { VMixPollingTimer } from '../vMixPollingTimer'
 
 describe('VMixPollingTimer', () => {
 	beforeEach(() => {
