+++ packages/timeline-state-resolver/src/integrations/vmix/vMixStateDiffer.ts
