+++ packages/timeline-state-resolver/src/integrations/vmix/vMixStateSynchronizer.ts
@@ -1,4 +1,4 @@
-import { VMixInput, VMixState, VMixStateExtended } from './WMixStateDiffer'
+import { VMixInput, VMixState, VMixStateExtended } from './vMixStateDiffer'
 import { EnforceableVMixInputStateKeys } from '.'
 import { VMixInputOverlays, VMixTransform } from 'timeline-state-resolver-types'
