+++ packages/timeline-state-resolver/src/integrations/vmix/vMixTimelineStateConverter.ts
@@ -10,7 +10,7 @@ import {
 	VMixTransition,
 	VMixTransitionType,
 } from 'timeline-state-resolver-types'
-import { TSR_INPUT_PREFIX, VMixInput, VMixInputAudio, VMixState, VMixStateExtended } from './WMixStateDiffer'
+import { TSR_INPUT_PREFIX, VMixInput, VMixInputAudio, VMixState, VMixStateExtended } from './vMixStateDiffer'
 import * as deepMerge from 'deepmerge'
 import _ = require('underscore')
